remove logs cte from user query
The original thought here was to alias the logs table with a blanket
query that added a limit on the number of logs to scan so we don't have
a possibility of a client requesting a full table scan.

This is a buggy way to solve the problem because the CTE limit will be
applied alongside a user limit which means that the user is unable to
query the full logs table.