add NOT MATERIALIZED to CTE queries to enable push down preds
This comes up when you have a query that references a CTE more than
once. In this case, the outer query may have predicates that the inner
CTE could use. An example is if you join the transfer table to itself
and you would like to add an address predicate. In this case, the
address predicate should be pushed down to the inner CTE query so that
we can make use of the (topics[1], address) index. But since the inner
CTE is referenced more than once, PG will MATERIALIZE the inner CTE and
therefore not accept the outer queries predicates.

See this [1] section of the PG docs for more detail

[1]: https://www.postgresql.org/docs/current/queries-with.html#QUERIES-WITH-CTE-MATERIALIZATION