relax select item re-writing
The previous code would attempt to rewrite select items and binary
expressions in such a way that if any identifier in an expression wasn't
able to be validated against the set of known Events then an error would
be returned. This rigid model broke down when using named expressions
for select items, for example: select foo as f, because the data model didn't
assume that f could be an alias for foo.

Instead, rewriting makes a best effort and doesn't return an error.

If there is a problem with the query, postgres will catch it and provide
a good error message. This makes our code much simpler.