add account based rate limiter
An account can have many api keys. Each account has a rate (per second)
and a timeout. Each api key has a list of valid origins. When a request
is made to the api, we load the account limit info from a hashmap. If
the api key's origins are not empty, we check the request host header to
see if it matches one of the valid origins. This isn't a perfect
security measure since the host header can be spoofed. However, it is
hard to spoof it in the browser. If the host header check passes, we
then check the account's rate limiter using the requester's IP address.
This allows IS customer's to use the SQL API as their API and we end up
rate limiting our customer's customer to X requests per second where X
is the account's rate limit.

This change does not limit the number of concurrent live queries. I will
address this in a later commit.