improve string literal fix for sql api
The fix in f144d5d95aac40897cd72bf6178b16e074ad666d wasn't quite
correct. We need to do the "last-mile" decoding for string.
Specifically, we wait to convert from abi encoded bytes to a string till
the projection part of the query.

I had to deal with the alloy crate and how it abi encodes strings. It
wanted to encode the string as a tuple. This means that the first 32
bytes are the position of the dynamic content. Since we only want the
length followed by the bytes of the string (padded to 32) I simply
remove the first 32 bytes of result.