add loading gif. refactor response div
the refactor introduces an idea where: examples, history, table data,
etc all live in the results. Anytime that an event would like to change
the results, it will call resetRestults which will hide all children
divs and then whatever new child is meant to be shown will have its
display set.

The result is that you don't have to know what to hide and you only have
to know what to show.