fix chain id in example queries
I recently made a change to the api.indexsupply.net ALB that
removed 8453 as the default chain when no indexed chain
could be found using the `?chain` query parameter.

I should also fix the API so that it returns a more helpful error
message when it receives a request for an unkown chain.