stop double logging api key in http requests
We record the api key in the limit trait and the key trait impls. I
didn't realize that recording twice would cause the data to be
duplicated in the span. I figure you create a span with an empty key and
then recording it twice would effectively overwrite the field in the
span.