tmp for for panic in sql api
When we receive a request for a chain that we haven't indexed we need to
provide a good error message. Right now this query will panic because it
finds no rows given the chain predicate.

This is a stop gap to at least not panic and it will return no data and
the block height will be 0.