expose logs table to user queries
Prior to this commit, the API required a set of event signatures
(although not explicitly) and the query's relations would be validated
against the names of the events provided in the event signatures.

This commit relaxes this check. We make a special case for the 'logs'
relation and selected columns are added to that relation's selected
fields.