stop writing processed queries to analytics table
I'd like to get to a place where we can simply write a serialized
request to the user_queries table. This can even happen as a middleware.
We can reproduce the re-written and generated queries at read
time.

One motivator for this change is to pull out the logging from the query
function. This will make the interfaces simpler.

The larger motivation for this is to re-use the query function for
webhooks.