improve user request query logs
I've been wanting to make this better for a while and I think I've
finally found an approach that I don't absolutely hate.

A couple of table stakes for logging data from a user's request:

1. It doesn't impact the user query. Either latency or errors
2. It does capture the events, query, and how long it took and the
   overall status of the request. We should always be "ahead" of a
   user's proor expereince. They should never have to give us info in order
   for us to debug the problem.
3. It doesn't make a mess of the code

I _think_ this patch accomplishes these things.

The order of the ServiceBuilder layers matters. I think I've got it
correct. Specifically I want to make sure that timeouts (caused by the
limit layer) are captured in the user_queries table with the 408 status.
In my local testing this seems to be the case with the current code.