fix fatal account_limits server crash
The tokio task construction was incorrect. Prior to this commit, if
there was a panic while loading account limits, the entire server
process would crash. If the account_limits panic was persistent this
would cause the server to become totally unavailable.

I found this out because if you delete a plan_change from the fe
database without deleting the account's api_keys then the account_limits
view would return a null value and the be's account_limits code was
unable to handle that and would panic.