check logs bloom for 0 logs
There is a bit of a race condition when we download logs on the head of
the chain. We ask for the latest block. We then ask for the logs and
finally the block and its transactions. It's possible that these
requests are distributed across inconsistent nodes on the backend. This
means we find out about block k but when request logs the node servicing
the request only knows about k-1. Eth clients will not report an error
when requesting logs beyond their chain height. One idea to mitigate
this issue is to check the block header for the logs_bloom when we have
0 logs. If the block header has a logs_bloom value then we likely have
hit this case. We return an error and then try again.