only consider rows that have at least the known size of the sig
There are sure to be collisions on the first 4 bytes of the input data
so we can narrow the results by ensuring that the input is at least the
static size of the type. Obviously the input data may be much larger for
dynamically sized payloads so this is not perfect.